externalize xpi download to template
Add-Ons are normally distributed in xpi archives via the addon web page
at addons.mozilla.org. Most packages use these as source as no better
option is available. The process is always the same and can thus easily
be captured by a template.

The only parameter is the extension name to compose the download URL.
This causes some packages to slightly adjust their name, which is
actually a good thing as they now exactly echo the name the original
author chose.