let makepkg handle extraction
The file format of Mozilla's xpis is unknown to makepkg, so it will not
extract them automatically, making necessary the convoluted
externalization of an extraction routine because pkgver needs
information from within the add-on but is always run before prepare.

But xpi is in fact just a renamed zip archive, so simply changing the
name makes makepkg handle the extraction. The archive is then removed
from the $srcdir in the now freed prepare function to keep it from being
installed by package functions that simply install everything they find.

This also removes a directory layer, further simplifying packaging in
most cases.