externalize version range to function
The resulting function is easier to export to a template to also use it
with single targeted Add-Ons.

Further externalization, i.e. also moving the assignment to $depends to
the function, is not possible, as this would keep makepkg -S from
recognizing the dependencies, as it only looks for these in immediate
package functions.