correct package relations
Packages build from Git repositories should usually provide and conflict
with their non-git variant. This now also holds true for the split
packages created for each target application.

Unfortunately, the parsing process of makepkg -S does not understand
conditional blocks, so these two arrays are also set up the same way
with packages that are not built without Git at all, making them provide
and conflict with themselves. This is ugly, but can be accepted, as it
does not break any actual functionality.