state relations to existing packages
Variants of this package, installing only for a single target
application, already exists under the names firefox-gnome-keyring and
thunderbird-gnome-keyring. Note the difference in the missing
"extension". There are also Git and precompiled binary variants.

All of them package a different extension that provides the same
functionality. Activating both at the same time is very likely to
confuse the target application, passwords ending up in random places, so
mozilla-extension-gnome-keyring-git conflicts with all of them.