simplify distribution
Now that the -b flag to makepkg-expanded is applicable to directories
holding nothing but the PKGBUILD, it can be used instead of a custom
distribution script.

This essentially reverts the last commit to "provide a distribution
helper" apart from the change to .gitignore, as it is not the
distribution helper but makepkg-expanded that leaves the files to be
ignored.