switch to fork to restore Firefox compatibility
The maintainer of the original source, swick, has not yet merged the
pull request for Firefox 40 while Firefox 41 is already out. The Add-On
seems to work mostly fine with the current version also, so switching to
a fork that raising the maxVersion to a ridiculously high 50 should be
fine until the original upstream updates.