rename targets template
First, this avoids the erroneous assumption it would be used to install
files, as its name started with "install-" before, making it
syntactically similar to the install-all template.

In addition, this frees the targets template from the restriction of
only relying upon the install.rdf file, so that other sources can also
be queried for compatible target application version information.