move compatible version query to source templates
This paves the way to query not only the install.rdf file, but also more
precise or up-to-date locations for target application version
restrictions. The availability of this information usually depends on
the source format of the Add-On, so it is now up to the templates
pulling in the sources to provide means of access to this information.