version Firefox dependency
For this to be done here analogous to how the package-multi and
package-single templates already handle it, the query-version originally
imported from source-git-plain via source-git-json has to be redirected
to the information in the JSON file.

This is not changed in source-git-json because the JSON files are no
required standard, like install.rdf, which are a mandatory part of every
extension, but a utility some developers collect deployment information
in to ease their own work. Thus, the hierarchy therein can change from
project to project, rendering a universal compatible version extraction
process to be mostly heuristic guesswork.

Some extensions, like gnotifier, do not include any compatibility
information in the JSON, but leave it in the install.rdf file. So
source-git-json is left better off parsing a, possibly generated,
install.rdf by default.