adapt to new directory structure
The relevant sources were recently moved into a dedicated directory.
The package template expects everything it should install directly in
the $srcdir. To restore compilation, the contents of the subdirectory
have to be moved down, similar to how it is done with the git repository
directory in the source-git-plain template.

In turn, the new hierarchy makes it easier to then eliminate files not
to be installed, as they now are mostly collected in directories.