remove nonexistent package relations
They never existed in AUR4. Enough time has passed that any leftover
installation should be outdated and incompatible so that they are not
relevant for an update any longer.

This only affects relations to no longer preset packages, which will not
disturb existing installations, so they do not need to be triggered to
update with a pkgrel increment.