derive URL from extension name
The description page for the Add-On follows a similar naming scheme as
the download link, so it can be automatically composed, too.

Only doing so when the $url is not yet set takes into account Add-Ons
that have a dedicated web page. The conditional would not be necessary
as long as the $url would be overwritten after template inclusion, which
is usually not the case: The template is semantically more suited to be
placed amongst the part of the PKGBUILD where prepare, build, package
and related functions are collected while the $url is set as part of an
earlier block defining variables.